requirements.txt
psycopg==3.1.18
pytest==8.1.1
pytest-cov==4.1.0
pytest-freezegun>=0.4.2
pytest-postgresql==5.1.1
pytest-xdist==3.6.1
SQLAlchemy==2.0.28
GeoAlchemy2==0.15.1
pyproj==3.6.1
requests==2.32.3
pytz>=2024.1
flask==3.0.3
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.1.1
python-dateutil==2.9.0
shapely==2.0.5
requests-mock==1.12.1